separate OS-es code
